+++ SwiftGrapher/Cells/EquationCellView.swift
