+++ SwiftGrapher/Helpers/EquationCalculationModel.swift
