+++ SwiftGrapher/Helpers/Logger.swift
