+++ SwiftGrapher/ViewControllers/SidebarViewController.swift
