+++ SwiftGrapher/Views/TransformManager.swift
